Fix domain lookup
