+++ server/middleware/configureEstablishment.js
@@ -1,5 +1,4 @@
 const { v4: uuid } = require('uuid');
-const { pathOr } = require('ramda');
 const {
   getEstablishmentId,
   getEstablishmentPersonalisation,
@@ -14,14 +13,15 @@ const configureEstablishment = () => (req, res, next) => {
       req.session.establishmentPersonalisationEnabled === undefined)
   ) {
     const replaceUrl = /-prisoner-content-hub.*$/g;
-    const establishmentName = pathOr('wayland', ['headers', 'host'], req)
+
+    const establishmentName = (req.headers?.host || 'wayland')
       .split('.')[0]
       .replace(replaceUrl, '');
 
     const establishmentId = getEstablishmentId(establishmentName);
 
     req.session.id = uuid();
-    if (typeof establishmentId === 'undefined') {
+    if (typeof establishmentId !== 'undefined') {
       req.session.establishmentId = establishmentId;
       req.session.establishmentName = establishmentName;
       req.session.establishmentPersonalisationEnabled = getEstablishmentPersonalisation(
